+++ tests/testthat/test-dataframe.R
@@ -617,9 +617,9 @@ test_that("drop_nulls", {
   expect_equal(pl$DataFrame(tmp)$drop_nulls("hp")$height, 32, ignore_attr = TRUE)
   expect_equal(pl$DataFrame(tmp)$drop_nulls(c("mpg", "hp"))$height, 29, ignore_attr = TRUE)
 
-  expect_identical(
-    result(pl$DataFrame(mtcars)$drop_nulls("bad column name")$height)$err$contexts(),
-    list(PolarsError = "not found: bad column name")
+  expect_error(
+    pl$DataFrame(mtcars)$drop_nulls("bad column name")$height,
+    "not found: unable to find column \"bad column name\""
   )
 })